+++ CHANGELOG.md
@@ -7,4 +7,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - Add implementation of MQTT client with optional delegation of Wi-Fi setup and monitoring.
+- Ability to specify a global prefix for the topic.
 - Add documentation for the MQTT client.